:recycle: refactor(tofhir-server-common): Move ToFhirRestCall of tofhir-log-server and tofhir-server.
