+++ tofhir-server/src/main/scala/io/tofhir/server/endpoint/ConceptMapEndpoint.scala
@@ -6,11 +6,11 @@ import akka.http.scaladsl.server.Route
 import com.typesafe.scalalogging.LazyLogging
 import io.tofhir.engine.Execution.actorSystem.dispatcher
 import io.tofhir.engine.config.ToFhirEngineConfig
+import io.tofhir.server.common.model.ToFhirRestCall
 import io.tofhir.server.endpoint.ConceptMapEndpoint.SEGMENT_CONCEPT_MAPS
 import io.tofhir.server.endpoint.TerminologyServiceManagerEndpoint._
 import io.tofhir.server.model.Json4sSupport._
 import io.tofhir.server.model.TerminologySystem.TerminologyConceptMap
-import io.tofhir.server.model.ToFhirRestCall
 import io.tofhir.server.service.ConceptMapService
 
 class ConceptMapEndpoint(toFhirEngineConfig: ToFhirEngineConfig) extends LazyLogging {
