+++ tofhir-server/src/main/scala/io/tofhir/server/endpoint/FhirPathFunctionsEndpoint.scala
@@ -3,9 +3,9 @@ package io.tofhir.server.endpoint
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.Route
 import com.typesafe.scalalogging.LazyLogging
+import io.tofhir.server.common.model.ToFhirRestCall
 import io.tofhir.server.endpoint.FhirPathFunctionsEndpoint.SEGMENT_FHIR_PATH_FUNCTIONS
 import io.tofhir.server.model.Json4sSupport._
-import io.tofhir.server.model.ToFhirRestCall
 import io.tofhir.server.service.FhirPathFunctionsService
 
 /**
