+++ tofhir-server/src/main/scala/io/tofhir/server/endpoint/ProjectEndpoint.scala
@@ -5,9 +5,10 @@ import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.Route
 import com.typesafe.scalalogging.LazyLogging
 import io.tofhir.engine.Execution.actorSystem.dispatcher
+import io.tofhir.server.common.model.ToFhirRestCall
 import io.tofhir.server.endpoint.ProjectEndpoint.SEGMENT_PROJECTS
 import io.tofhir.server.model.Json4sSupport._
-import io.tofhir.server.model.{Project, ToFhirRestCall}
+import io.tofhir.server.model.Project
 import io.tofhir.server.service.ProjectService
 import io.tofhir.server.service.job.IJobRepository
 import io.tofhir.server.service.mapping.IMappingRepository
