+++ tofhir-server/src/main/scala/io/tofhir/server/interceptor/IErrorHandler.scala
@@ -2,7 +2,8 @@ package io.tofhir.server.interceptor
 
 import akka.http.scaladsl.server.{Directives, ExceptionHandler}
 import com.typesafe.scalalogging.LazyLogging
-import io.tofhir.server.model.{InternalError, ToFhirError, ToFhirRestCall}
+import io.tofhir.server.common.model.ToFhirRestCall
+import io.tofhir.server.model.{InternalError, ToFhirError}
 
 trait IErrorHandler extends LazyLogging {
